+++ mldsa_openssl_interop_test.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-//go:build opensslinterop
+//go:build opensslinterop && !nomldsa
 
 package tls
