+++ mldsa_test_fixtures_test.go
@@ -1,3 +1,5 @@
+//go:build !nomldsa
+
 package tls
 
 const testRootCertPEM = `
